requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
torch>=2.0.0, <3.0.0
torchvision>=0.15.0, <1.0.0
pillow>=10.0.0
numpy>=1.22
pytest>=7.0.0
httpx>=0.23.0
python-multipart>=0.0.7
requests>=2.28.0 